+++ aoc-2020/day10/day10.h
@@ -41,7 +41,7 @@ public:
 
   // Part2
   void solver_part2(void);
-  int eval_number_combis(void);
+  long long eval_number_combis(void);
 
 
   // Common
